requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-translate
numpy
transformers
torch
ndjson
scikit-learn